#167ベストアンサー機能

+++ src/components/pages/Post.vue
@@ -1,9 +1,11 @@
 <template>
   <div>
-    <h2>質問{{postId}}のページ</h2>
+    <h2>質問</h2>
+    <button @click="addNotification">テスト用ボタン（ここを押すと通知情報が追加されます）</button>
 
     <hr>
 
+    <!-- 質問 -->
     <div class='content-box'>
       <div>
         投稿者：
@@ -38,77 +40,154 @@
 
     <hr>
 
-    <div class='post-form'>
-      <div>
-          <label for='answer'>*回答内容</label>
-          <br>
-          <textarea id='answer' cols='30' rows='10' v-model='answer'></textarea>
+    <!-- ベストアンサーが選ばれたら回答エリアを消す -->
+    <template v-if='!post.document.fields.isResolved.booleanValue && (thread === null ? true : (typeof(thread.isResolved) === "undefined") ? true : !thread.isResolved.booleanValue)'>
+      <!-- 回答エリア -->
+      <div class='post-form'>
+        <div>
+            <label for='answer'>*回答内容</label>
+            <br>
+            <textarea id='answer' cols='30' rows='10' v-model='answer'></textarea>
+        </div>
       </div>
-    </div>
 
-    <br>
-    <button @click="addNotification">テスト用ボタン（ここを押すと通知情報が追加されます）</button>
-    <hr>
+      <hr>
 
-    <button @click='addAnswer'>回答を送信</button>
+      <button @click='addAnswer'>回答を送信</button>
+    </template>
 
     <br><br>
     <br><br>
 
+    <!-- 回答がある場合 -->
     <template v-if='this.isAnswered'>
       <h3>回答{{thread.answers.arrayValue.values.length}}件</h3>
 
       <hr>
 
-      <div v-for='(answer, index) in thread.answers.arrayValue.values' :key='answer.mapValue.fields.answerId.stringValue' class='content-box'>
-        <h3>回答</h3>
-        <div>
-          {{answer.mapValue.fields.userName.stringValue}}さん
-        </div>
-        <div>
-          {{answer.mapValue.fields.answer.stringValue}}
-        </div>
-        <div>
-          {{answer.mapValue.fields.created_at.timestampValue | dateFormat}}
-        </div>
-        <template v-if='uid == answer.mapValue.fields.uid.stringValue'>
-          <button @click='deleteAnswer(answer.mapValue.fields.answerId.stringValue)'>削除</button>
-        </template>
-        <hr>
-
-        <div v-for='comment in answer.mapValue.fields.comments.arrayValue.values' :key='comment.mapValue.fields.commentId.stringValue'>
-          <h3>コメント</h3>
+      <!-- ベストアンサー -->
+      <div v-for='(answer, index) in thread.answers.arrayValue.values' :key='answer.mapValue.fields.answerId.stringValue + "best"' class='content-box'>
+        <template v-if='answer.mapValue.fields.isBestAnswer.booleanValue'>
+          <h3>ベストアンサーに選ばれた回答</h3>
           <div>
-            {{comment.mapValue.fields.userName.stringValue}}さん
+            {{answer.mapValue.fields.userName.stringValue}}さん
           </div>
           <div>
-            {{comment.mapValue.fields.comment.stringValue}}
+            {{answer.mapValue.fields.answer.stringValue}}
           </div>
           <div>
-            {{comment.mapValue.fields.created_at.timestampValue | dateFormat}}
+            {{answer.mapValue.fields.created_at.timestampValue | dateFormat}}
           </div>
-          <template v-if='uid == comment.mapValue.fields.uid.stringValue'>
-            <button @click='deleteComment(answer.mapValue.fields.answerId.stringValue, comment.mapValue.fields.commentId.stringValue)'>削除</button>
-          </template>
-          <hr>
-        </div>
-
-        <template v-if='isDisplayCommentArea[index].value'>
-          <div class='post-form'>
+          <template v-if='uid == answer.mapValue.fields.uid.stringValue'>
             <div>
+              <button @click='deleteAnswer(answer.mapValue.fields.answerId.stringValue, answer.mapValue.fields.isBestAnswer.booleanValue)'>削除</button>
+            </div>
+          </template>
+
+          <hr>
+
+          <!-- コメントエリア -->
+          <div v-for='comment in answer.mapValue.fields.comments.arrayValue.values' :key='comment.mapValue.fields.commentId.stringValue + "best"'>
+            <h3>コメント</h3>
+            <div>
+              {{comment.mapValue.fields.userName.stringValue}}さん
+            </div>
+            <div>
+              {{comment.mapValue.fields.comment.stringValue}}
+            </div>
+            <div>
+              {{comment.mapValue.fields.created_at.timestampValue | dateFormat}}
+            </div>
+            <template v-if='uid == comment.mapValue.fields.uid.stringValue'>
+              <button @click='deleteComment(answer.mapValue.fields.answerId.stringValue, comment.mapValue.fields.commentId.stringValue)'>削除</button>
+            </template>
+            <hr>
+          </div>
+
+          <template v-if='isDisplayCommentArea[index].value'>
+            <div class='post-form'>
+              <div>
                 <label for='comment'>*コメント内容</label>
                 <br>
                 <textarea id='comment' cols='30' rows='10' v-model='comment[index].value'></textarea>
+              </div>
             </div>
+
+            <button @click='addComment(answer.mapValue.fields.answerId.stringValue, index)'>コメントを送信</button>
+          </template>
+          <template v-else>
+            <button @click='displayCommentArea(index)'>コメントする</button>
+          </template>
+        </template>
+      </div>
+
+      <!-- ベストアンサーと回答を区切るための<hr> -->
+      <template v-if='!(!post.document.fields.isResolved.booleanValue && (thread === null ? true : (typeof(thread.isResolved) === "undefined") ? true : !thread.isResolved.booleanValue))'>
+        <hr>
+      </template>
+
+      <!-- 回答 -->
+      <div v-for='(answer, index) in thread.answers.arrayValue.values' :key='answer.mapValue.fields.answerId.stringValue' class='content-box'>
+        <template v-if='!answer.mapValue.fields.isBestAnswer.booleanValue'>
+          <h3>回答</h3>
+          <div>
+            {{answer.mapValue.fields.userName.stringValue}}さん
+          </div>
+          <div>
+            {{answer.mapValue.fields.answer.stringValue}}
+          </div>
+          <div>
+            {{answer.mapValue.fields.created_at.timestampValue | dateFormat}}
+          </div>
+          <template v-if='!answer.mapValue.fields.isBestAnswer.booleanValue && uid == post.document.fields.uid.stringValue && !thread.isResolved.booleanValue'>
+            <div>
+              <button @click='updateBestAnswer(answer.mapValue.fields.answerId.stringValue)'>ベストアンサーにする</button>
+            </div>
+          </template>
+
+          <template v-if='uid == answer.mapValue.fields.uid.stringValue'>
+            <div>
+              <button @click='deleteAnswer(answer.mapValue.fields.answerId.stringValue)'>削除</button>
+            </div>
+          </template>
+
+          <hr>
+
+          <!-- コメントエリア -->
+          <div v-for='comment in answer.mapValue.fields.comments.arrayValue.values' :key='comment.mapValue.fields.commentId.stringValue'>
+            <h3>コメント</h3>
+            <div>
+              {{comment.mapValue.fields.userName.stringValue}}さん
+            </div>
+            <div>
+              {{comment.mapValue.fields.comment.stringValue}}
+            </div>
+            <div>
+              {{comment.mapValue.fields.created_at.timestampValue | dateFormat}}
+            </div>
+            <template v-if='uid == comment.mapValue.fields.uid.stringValue'>
+              <button @click='deleteComment(answer.mapValue.fields.answerId.stringValue, comment.mapValue.fields.commentId.stringValue)'>削除</button>
+            </template>
+            <hr>
           </div>
 
-          <button @click='addComment(answer.mapValue.fields.answerId.stringValue, index)'>コメントを送信</button>
-        </template>
-        <template v-else>
-          <button @click='displayCommentArea(index)'>コメントする</button>
-        </template>
+          <template v-if='isDisplayCommentArea[index].value'>
+            <div class='post-form'>
+              <div>
+                <label for='comment'>*コメント内容</label>
+                <br>
+                <textarea id='comment' cols='30' rows='10' v-model='comment[index].value'></textarea>
+              </div>
+            </div>
 
-        <hr>
+            <button @click='addComment(answer.mapValue.fields.answerId.stringValue, index)'>コメントを送信</button>
+          </template>
+          <template v-else>
+            <button @click='displayCommentArea(index)'>コメントする</button>
+          </template>
+
+          <hr>
+        </template>
       </div>
     </template>
 
@@ -180,6 +259,7 @@ export default {
       this.$store.dispatch("notification/addNotification", notificationData);
     },
     addAnswer() {
+      // 回答が存在しない場合、回答なし --> 回答ありにする
       if (!this.isAnswered) {
         this.$store.dispatch('post/updateIsAnswered', {
           postId: this.postId,
@@ -195,6 +275,7 @@ export default {
           fields: {
             answer: { stringValue: this.answer },
             answerId: { stringValue: new Date().getTime().toString(16) + Math.floor(1000 * Math.random()).toString(16) },
+            isBestAnswer: { booleanValue: false },
             uid: { stringValue: this.$store.getters.uid },
             userName: { stringValue: this.$store.getters.userName },
             created_at: { timestampValue: new Date().toISOString() },
@@ -205,6 +286,7 @@ export default {
       }
 
       if (this.threadExists) {
+        // スレッドが存在する場合、スレッドに回答を追加し、回答ありに設定
         this.$store.dispatch('thread/addThread', {
           postId: this.postId,
           answer: answer,
@@ -215,6 +297,7 @@ export default {
           }
         });
       } else {
+        // スレッドが存在しない場合、スレッドを作成し、スレッドあり・回答ありに設定
         this.$store.dispatch('thread/createThread', {
           postId: this.postId,
           fields: {
@@ -224,20 +307,22 @@ export default {
                   answer
                 ]
               }
-            }
+            },
+            isResolved: { booleanValue: false },
+            created_at: { timestampValue: new Date().toISOString() },
           },
         }).then((response) => {
           this.threadExists = true;
           if (response == 'OK') {
             this.isAnswered = true;
           } else if (response == 'ALREADY_EXISTS') {
+            // スレッドが存在する場合、スレッドに回答を追加し、回答ありに設定
             this.$store.dispatch('thread/addThread', {
               postId: this.postId,
               answer: answer,
               type: 'answer',
             }).then((response) => {
               if (response == 'OK') {
-                console.log('ALREADY_EXISTS');
                 this.isAnswered = true;
               }
             });
@@ -273,26 +358,48 @@ export default {
     displayCommentArea(index) {
       this.isDisplayCommentArea[index].value = true;
     },
-    deleteAnswer(answerId) {
-      this.$store.dispatch('thread/deleteAnswer', {
-        postId: this.postId,
-        answerId,
-      }).then(() => {
-        if (this.$store.getters.thread.answers.arrayValue.values.length == 0) {
-          this.isAnswered = false;
-          this.$store.dispatch('post/updateIsAnswered', {
-            postId: this.postId,
-            isAnswered: false
-          });
-        }
-      });
+    deleteAnswer(answerId, isBestAnswer) {
+      if (confirm('本当にこの回答を削除しますか？')) {
+        this.$store.dispatch('thread/deleteAnswer', {
+          postId: this.postId,
+          answerId,
+        }).then(() => {
+          // 全ての回答が削除された場合、回答あり --> 回答なし
+          if (this.$store.getters.thread.answers.arrayValue.values.length == 0) {
+            this.isAnswered = false;
+            this.$store.dispatch('post/updateIsAnswered', {
+              postId: this.postId,
+              isAnswered: false
+            });
+          }
+          // ベストアンサーが削除された場合、解決済み --> 未解決にする
+          if (isBestAnswer) {
+            this.$store.dispatch('thread/updateBestAnswer', {
+              postId: this.postId,
+              answerId,
+              isResolved: false
+            });
+          }
+        });
+      }
     },
     deleteComment(answerId, commentId) {
-      this.$store.dispatch('thread/deleteComment', {
-        postId: this.postId,
-        answerId,
-        commentId,
-      })
+      if (confirm('本当にこのコメントを削除しますか？')) {
+        this.$store.dispatch('thread/deleteComment', {
+          postId: this.postId,
+          answerId,
+          commentId,
+        })
+      }
+    },
+    updateBestAnswer(answerId) {
+      if (confirm('この回答をベストアンサーにしますか？')) {
+        this.$store.dispatch('thread/updateBestAnswer', {
+          postId: this.postId,
+          answerId,
+          isResolved: true
+        });
+      }
     }
   },
   created() {
